+++ server/src/constants/queue.constants.js
@@ -1,4 +1,4 @@
-const VIDEO_QUEUE_EVENTS = {
+export const VIDEO_QUEUE_EVENTS = {
   VIDEO_UPLOADED: "video.uploaded",
   VIDEO_PROCESSING: "video.processing",
   VIDEO_PROCESSED: "video.processed",
@@ -7,13 +7,11 @@ const VIDEO_QUEUE_EVENTS = {
   VIDEO_HLS_CONVERTED: "video.hls.converted",
 };
 
-const NOTIFY_EVENTS = {
+export const NOTIFY_EVENTS = {
   NOTIFY_VIDEO_HLS_CONVERTED: "notify.video.hls.converted",
 };
 
-const ALL_EVENTS = {
+export const ALL_EVENTS = {
   ...VIDEO_QUEUE_EVENTS,
   ...NOTIFY_EVENTS,
 };
-
-module.exports = { VIDEO_QUEUE_EVENTS, NOTIFY_EVENTS, ALL_EVENTS };
