+++ server/src/middlewares/index.js
@@ -1,0 +1,1 @@
+export { default as uploadMiddleware } from "./upload.middleware.js";
